+++ packages/sznm_lints/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.0
 
- - **FEAT**: add sznm_lints package. ([461f8d6c](https://github.com/sozonome/sznm_dart_packages/commit/461f8d6c54e1efddd0ab049e8621c5abf25879ae))
+ - **FEAT**: add sznm_lints package. ([461f8d6c](https://github.com/agustinusnathaniel/sznm_dart_packages/commit/461f8d6c54e1efddd0ab049e8621c5abf25879ae))
 
 ## 0.0.1